Moves the data model to the package org.orbisgis.orbistoolbox.model.
