+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/process/model/Values.java
@@ -17,41 +17,14 @@
  * For more information, please consult: <http://www.orbisgis.org/> or contact directly: info_at_orbisgis.org
  */
 
-package org.orbisgis.orbistoolbox.process;
+
+package org.orbisgis.orbistoolbox.process.model;
 
 /**
- * Simple value. The value is represented as a String.
- * This class comes from the OWS specification : ows:ValueType and ows:Value
+ * This class represent a value and is the super class of Range and Value.
  *
  * @author Sylvain PALOMINOS
  */
 
-public class Value extends Values {
-    /** String representation of the value. */
-    private double value;
-
-    /**
-     * Main constructor in which the value must be passed as a String.
-     * The value can not be null.
-     * @param value Not null value.
-     */
-    public Value(double value) {
-        this.value = value;
-    }
-
-    /**
-     * Returns the value.
-     * @return The value.
-     */
-    public double getValue() {
-        return value;
-    }
-
-    /**
-     * Sets the the value. The value given as parameter can not be null.
-     * @param value Not null value.
-     */
-    public void setValue(double value) {
-        this.value = value;
-    }
+public abstract class Values {
 }
