+++ bundles/orbistoolbox/src/test/java/org/orbisgis/orbistoolbox/process/model/RawDataTest.java
@@ -17,9 +17,11 @@
  * For more information, please consult: <http://www.orbisgis.org/> or contact directly: info_at_orbisgis.org
  */
 
-package org.orbisgis.orbistoolbox.process;
+package org.orbisgis.orbistoolbox.process.model;
 
 import org.junit.Test;
+import org.orbisgis.orbistoolbox.process.model.Format;
+import org.orbisgis.orbistoolbox.process.model.RawData;
 
 import java.net.URISyntaxException;
 import java.net.URI;
